--- service/src/main/java/org/databiosphere/workspacedataservice/CollectionInitializer.java
@@ -1,25 +1,0 @@
-package org.databiosphere.workspacedataservice;
-
-import org.databiosphere.workspacedataservice.annotations.DeploymentMode.DataPlane;
-import org.jetbrains.annotations.NotNull;
-import org.springframework.context.ApplicationListener;
-import org.springframework.context.annotation.Profile;
-import org.springframework.context.event.ContextRefreshedEvent;
-import org.springframework.stereotype.Component;
-
-@Component
-@DataPlane
-@Profile({"!local"})
-public class CollectionInitializer implements ApplicationListener<ContextRefreshedEvent> {
-
-  private final CollectionInitializerBean collectionInitializerBean;
-
-  public CollectionInitializer(CollectionInitializerBean collectionInitializerBean) {
-    this.collectionInitializerBean = collectionInitializerBean;
-  }
-
-  @Override
-  public void onApplicationEvent(@NotNull ContextRefreshedEvent event) {
-    collectionInitializerBean.initializeCollection();
-  }
-}
